Crate iota_json

Macros

call_arg

call_args

type_args

Structs

IotaJsonValue

IotaJsonValueError

Enums

IotaJsonValueErrorKind

A list of error categories encountered when parsing numbers.

MoveTypeLayout

ResolvedCallArg

Functions

check_valid_homogeneous

Check via BFS The invariant is that all types at a given level must be the same or be empty, and all must be valid

is_receiving_argument

primitive_type

Checks if a give SignatureToken represents a primitive type and, if so, returns MoveTypeLayout for this type (if available). The reason we need to return both information about whether a SignatureToken represents a primitive and an Option representing MoveTypeLayout is that there can be signature tokens that represent primitives but that do not have corresponding MoveTypeLayout (e.g., SignatureToken::DatatypeInstantiation).

resolve_call_args

resolve_move_function_args

Resolve the JSON args of a function into the expected formats to make them usable by Move call This is because we have special types which we need to specify in other formats