iota_json

Function resolve_move_function_args

pub fn resolve_move_function_args(
    module: &CompiledModule,
    function: &Identifier,
    type_args: &[TypeTag],
    args: Vec<IotaMoveCallInputValue>,
    require_view_function: bool,
) -> Result<Vec<(ResolvedCallArg, SignatureToken)>, Error>

Resolve the JSON args of a function into the expected formats to make them usable by Move call. This is because we have special types which we need to specify in other formats. Additionally, it checks for #[view] attribute presence in function metadata if a view/non-view function is expected.