iota_json

Function resolve_move_function_args

pub fn resolve_move_function_args(
    package: &MovePackage,
    module_ident: Identifier,
    function: Identifier,
    type_args: &[TypeTag],
    combined_args_json: Vec<IotaJsonValue>,
) -> Result<Vec<(ResolvedCallArg, SignatureToken)>, Error>

Resolve the JSON args of a function into the expected formats to make them usable by Move call This is because we have special types which we need to specify in other formats